Fixed compilation warnings on mac os x.

+++ lib/map.c
@@ -20,7 +20,7 @@ void mocha_map_init(mocha_map* self, const mocha_object* args[], int count)
 
 const struct mocha_object* mocha_map_lookup(const mocha_map* self, const struct mocha_object* key)
 {
-	for (int i=0; i<self->count; i += 2) {
+	for (size_t i=0; i<self->count; i += 2) {
 		if (mocha_object_equal(self->objects[i], key)) {
 			return self->objects[i + 1];
 		}
